+++ test/test.js
@@ -10,6 +10,11 @@ import { RenderAsLayers } from '../lib/render/renderAsLayers.js'
 import { RenderAsPoster } from '../lib/render/renderAsPoster.js'
 import { GLAD } from '../lib/glad.js'
 
+// Import additional test files
+import '../test_swift_dependency.js'
+import '../test_js_dependency.js'
+import '../test_ts_dependency.js'
+
 const graph = new Graph()
 graph.rootNode = new Container(null, 'Farming')